+++ src/phoenix/trace/langchain/__init__.py
@@ -1,0 +1,3 @@
+from .tracer import OpenInferenceTracer
+
+__all__ = ["OpenInferenceTracer"]
