+++ src/phoenix/trace/llama_index/__init__.py
@@ -1,0 +1,3 @@
+from .callback import OpenInferenceTraceCallbackHandler
+
+__all__ = ["OpenInferenceTraceCallbackHandler"]
